+++ src/main/java/com/tjx/lew00305/slimstore/model/report/TransactionLineFlat.java
@@ -1,5 +1,7 @@
 package com.tjx.lew00305.slimstore.model.report;
 
+import java.math.BigDecimal;
+
 import com.tjx.lew00305.slimstore.model.entity.TransactionLine.TransactionLineType;
 
 import lombok.Data;
@@ -17,7 +19,7 @@ public class TransactionLineFlat {
     private String product;
     private TransactionLineType type;
     private Integer quantity;
-    private Float lineValue;
+    private BigDecimal lineValue;
     private Integer returnedQuantity;
     
 }
